add query builder module
