+++ tap_suiteql/client.py
@@ -15,9 +15,13 @@ class suiteqlStream(RESTStream):
     """suiteql stream class."""
 
     def __init__(
-        self, tap: Any, schema: dict = {"type": "object", "properties": {}}
+        self,
+        tap: Any,
+        schema: dict = {"type": "object", "properties": {}},
+        body_query: str = "",
     ) -> None:
         super().__init__(tap=tap, schema=schema)
+        self.body_query = body_query
 
     rest_method = "POST"
     stream_type = ""
@@ -30,7 +34,6 @@ class suiteqlStream(RESTStream):
 
     next_page_token_jsonpath = "$.links[?(@.rel == 'next')].href"
 
-    body_query = ""
     metadata_path = ""
     skip_attributes = []
 
@@ -141,7 +144,7 @@ class suiteqlStream(RESTStream):
         """
         start_date = self.config.get("start_date")
         bookmark_date = self.get_starting_timestamp(context)
-        current_body = QueryBuilder(self).query()
+        current_body = self.body_query
 
         if bookmark_date:
             start_date = bookmark_date.strftime("%Y-%m-%dT%H:%M:%S")
@@ -164,50 +167,3 @@ class suiteqlStream(RESTStream):
         """
 
         return row
-
-
-class QueryBuilder:
-    SELECT_STATEMENT = "select "
-    WHERE_STATEMENT = "where "
-
-    def __init__(self, stream: suiteqlStream):
-        self.stream: suiteqlStream = stream
-
-    def _get_column_select(self, schema: dict) -> List:
-        column_select = []
-        for attribute_name, properties in schema["properties"].items():
-            if properties.get("format") == "date-time":
-                column_select.append(
-                    f"""TO_CHAR({attribute_name}, 'YYYY-MM-DD\"T\"HH24:MI:SS') {attribute_name}"""
-                )
-            else:
-                column_select.append(attribute_name)
-        return column_select
-
-    def _query_builder(
-        self, schema: dict, replication_key: str, entity_name: str, stream_type: str
-    ) -> str:
-        from_statement = f"from {entity_name}"
-        where_clauses = ["1=1"]
-        column_select = self._get_column_select(schema)
-        if replication_key:
-            where_clauses.append(
-                f"{replication_key} >= TO_DATE(:{replication_key}, 'YYYY-MM-DD\"T\"HH24:MI:SS')"
-            )
-        if stream_type:
-            from_statement = f"from transaction"
-            where_clauses.append(f"type = '{stream_type}'")
-        self.SELECT_STATEMENT += ",".join(column_select)
-        self.WHERE_STATEMENT += " and ".join(where_clauses)
-        query = (
-            f"{self.SELECT_STATEMENT} {from_statement} {self.WHERE_STATEMENT}".strip()
-        )
-        return query
-
-    def query(self):
-        return self._query_builder(
-            self.stream.schema,
-            self.stream.replication_key,
-            self.stream.name,
-            self.stream.stream_type,
-        )
