+++ tap_suiteql/tests/test_query_builder.py
@@ -1,6 +1,6 @@
 from isort import stream
 from singer_sdk import typing as th
-from tap_suiteql.client import QueryBuilder
+from tap_suiteql.query_builder import QueryBuilder
 
 
 class DummyStream:
